requirements.txt
alembic>=1.14.1
backoff==2.2.1
bcrypt==4.1.2
dramatiq==1.17.1
python-box
clickhouse-sqlalchemy==0.3.2
fastapi==0.109.0
geopy==2.4.1
httpx==0.28.1
lancedb==0.18.0
lxml==5.3.0
passlib==1.7.4
pydantic==2.5.3
python-jose==3.3.0
python-multipart==0.0.6
redis==5.2.1
requests==2.32.3
requests-cache==1.2.1
sentence-transformers==2.2.2
sqlalchemy==2.0.36
transformers==4.36.2
uvicorn==0.27.0
warcio==1.7.5